Очень длинный вызов метода, как результат outofmemoryerror - 70%, найденный ключ 30%
