--- scr/java/com/javarush/cryptoanalyser/input.txt
@@ -1,1 +1,0 @@
-привет мир!
