+++ scr/java/com/javarush/cryptoanalyser/library/input.txt
@@ -1,0 +1,1 @@
+Привет мой дорогой друг! Я давно о тебе думал.
