--- scr/java/com/javarush/cryptoanalyser/someText.txt
@@ -1,1 +1,0 @@
-уфнжкцгрнфб
